+++ api/tests/tests.go
@@ -185,7 +185,7 @@ func (th *testHarness) run(
 			wg.Add(1)
 			go func(x int, config gofig.Config) {
 				defer wg.Done()
-				server, err, errs := apiserver.Serve(config)
+				server, errs, err := apiserver.Serve(config)
 				if err != nil {
 					th.closeServers(t)
 					t.Fatal(err)
@@ -231,7 +231,7 @@ func (th *testHarness) run(
 				go func(test APITestFunc, x int, config gofig.Config) {
 
 					defer wg.Done()
-					server, err, errs := apiserver.Serve(config)
+					server, errs, err := apiserver.Serve(config)
 					if err != nil {
 						th.closeServers(t)
 						t.Fatal(err)
