+++ glide.lock
@@ -1,5 +1,5 @@
 hash: 0b1c2794d82975d611e966e8102fd95d1bee1ff1b81e2d592f26026e41c2d65e
-updated: 2016-05-26T16:59:16.095982253-05:00
+updated: 2016-05-26T20:02:41.60257948-05:00
 imports:
 - name: github.com/akutz/gofig
   version: 021bbefbfe234a3b72dc9137847483f5b997cbd0
@@ -81,7 +81,7 @@ imports:
   subpackages:
   - assert
 - name: golang.org/x/net
-  version: dfd9ed699ea197b58a32e0d1dcfa2be66f127ec4
+  version: 30db96677b74e24b967e23f911eb3364fc61a011
   subpackages:
   - context/ctxhttp
   - context
